fix build errors in bm and update bm impl of lccontdirect

+++ benchmarks/src/FullComplexityBenchmarks.cpp
@@ -9,6 +9,7 @@
  */
 #include <benchmark/benchmark.h>
 
+#include <cstddef>
 #include <memory>
 
 #include "io/outputWriter/XYZWriter.h"
@@ -31,7 +32,8 @@ void bmSimulationComplexityLinkedCell(benchmark::State& state) {
     LinkedCellContainer part_container({180., 90., 1.}, 3.0);
     ContainerRef particles(part_container);
 
-    CuboidGenerator generator({60.0, 60.0, 0.0}, {0.0, 0.0, 0.0}, {n, n, 1}, 1.0, 1.1225, 0.3, 5.0, 1.0);
+    CuboidGenerator generator({60.0, 60.0, 0.0}, {0.0, 0.0, 0.0}, {n, n, static_cast<size_t>(1)}, 1.0, 1.1225, 0.3, 5.0,
+                              1.0);
     SettingsParam settings;
     settings.delta_t = 0.0005;
     settings.start_time = 0;
@@ -69,7 +71,8 @@ void bmSimulationComplexityDirectSum(benchmark::State& state) {
     size_t n = state.range(0);
     SimpleContainer part_container;
     ContainerRef particles(part_container);
-    CuboidGenerator generator({60.0, 60.0, 0.0}, {0., 0., 0.}, {n, n, 1}, 1.0, 1.1225, 0.3, 5.0, 1.0);
+    CuboidGenerator generator({60.0, 60.0, 0.0}, {0., 0., 0.}, {n, n, static_cast<size_t>(1)}, 1.0, 1.1225, 0.3, 5.0,
+                              1.0);
     SettingsParam settings;
     settings.delta_t = 0.0005;
     settings.start_time = 0;
